src/requirements.txt
appdirs==1.4.0
boto==2.47.0
braintree==3.35.0
Django==1.10.5
django-allauth==0.30.0
django-crispy-forms==1.6.1
django-storages==1.5.2
numpy==1.12.1
oauthlib==2.0.1
olefile==0.44
packaging==16.8
pandas==0.19.2
Pillow==4.0.0
psycopg2==2.6.2
pyparsing==2.1.10
python-dateutil==2.6.0
python-openid==2.2.5
pytz==2017.2
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
